added mini console to list/create/delete announcements from command line

+++ lib/announcement_fu.rb
@@ -21,7 +21,7 @@ end
   path = File.join(File.dirname(__FILE__), 'app', dir)  
   $LOAD_PATH << path 
   ActiveSupport::Dependencies.load_paths << path 
-  ActiveSupport::Dependencies.load_once_paths.delete(path) 
+  ActiveSupport::Dependencies.load_once_paths.delete(path) if RAILS_ENV=="development"
 end 
 
 module ApplicationHelper
